Move Statement v1 APIs to separate directory hierarchy
