+++ go.mod
@@ -4,7 +4,7 @@ go 1.20
 
 require (
 	github.com/google/go-cmp v0.6.0
-	github.com/in-toto/attestation v0.1.1-0.20230828220013-11b7a1a4ca51
+	github.com/in-toto/attestation v0.1.1-0.20230908124000-cefdc8c70608
 	github.com/secure-systems-lab/go-securesystemslib v0.7.0
 	github.com/shibumi/go-pathspec v1.3.0
 	github.com/spf13/cobra v1.8.0
